Update README.md
The "Image Search Engine" project is a dynamic web application designed to provide users with a seamless and visually engaging experience in discovering high-quality images. Developed using a combination of HTML, CSS, and JavaScript, this project leverages the power of the Unsplash free API to fetch diverse and copyright-free images based on user-entered keywords.